src/requirements.txt
deepmerge == 1.1.0
docker == 6.1.3
pydantic == 2.3.0
pyyaml == 6.0.0
